participants/itd/archives/20260417162848-repo-url-changed/requirements.txt
alembic==1.4.3
blinker==1.4
Bootstrap-Flask==1.5
cffi==1.14.3
click==7.1.2
cryptography==2.9.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jwt==1.0.0
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2==2.8.6
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
WTForms==2.3.3
